Get from base task for dashboard by userId.

+++ Core.Application/Queries/Tasks/GetTasksByUserIdQuery.cs
@@ -10,9 +10,9 @@ namespace Core.Application.Queries
 {
     public class GetTasksForDashboardByUserIdQuery : IRequest<IEnumerable<TaskForDashboardDto>>
     {
-        public int UserId { get; }
+        public long UserId { get; }
 
-        public GetTasksForDashboardByUserIdQuery(int userId)
+        public GetTasksForDashboardByUserIdQuery(long userId)
         {
             UserId = userId;
         }
